+++ scraper/src/maps2zim/processor.py
@@ -230,7 +230,8 @@ class Processor:
                 secondary_color=self.zim_config.secondary_color,
                 zim_name=self.formatted_config.name,
                 center=(
-                    [context.default_view[0], context.default_view[1]]
+                    # maplibre center expects lon,lat format
+                    [context.default_view[1], context.default_view[0]]
                     if context.default_view
                     else None
                 ),