+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Tiles are not compressed by libzim (#47)
+- Default longitude and latitude handling is wrong (#43)
 
 ## [0.1.0] - 2026-03-06